fix(exec): recover precise confinement diagnostics and path hints
Recover the unfinished confinement lane, distinguish read and write evidence from generic refusals, and keep path-shaped patterns out of file advice. Preserve checker limits by moving pure classification into the analyzer and splitting seatbelt fixtures.
