+++ changelog.d/exec-refusal-names-the-path.fixed.md
@@ -1,0 +1,1 @@
+A seatbelt confinement refusal names a path only when stderr identifies it. `sandbox-exec` `file-read`/`file-write` evidence is distinct from a `prog: path: EPERM` line (class unknown). A named path is not proof a grant is missing (OS ACLs can deny a granted path). `nika explain NIKA-SEC-001` tells the author to check read/write authority and OS access; unknown stays unknown. A `unix://` URI on a denial line names the socket. The docker-report template does not promise Linux socket reachability.
