+++ crates/nika-check-analyzer/public-api.txt
@@ -239,9 +239,17 @@ pub fn nika_check_analyzer::edges::producer_ids(task: &nika_schema::raw::task::R
 pub fn nika_check_analyzer::edges::role_of_field(field: core::option::Option<&str>) -> nika_check_analyzer::edges::EdgeKind
 pub fn nika_check_analyzer::edges::task_refs_in_value(value: &serde_json::value::Value, out: &mut alloc::vec::Vec<(alloc::string::String, core::option::Option<alloc::string::String>)>)
 pub fn nika_check_analyzer::edges::witness_routes(edges: &[nika_check_analyzer::edges::Edge], nodes: usize, source: usize) -> alloc::vec::Vec<nika_check_analyzer::edges::Route>
+pub mod nika_check_analyzer::exec_read_path
+pub const nika_check_analyzer::exec_read_path::RULE: &str
+pub fn nika_check_analyzer::exec_read_path::scan(wf: &nika_schema::raw::workflow::RawWorkflow) -> alloc::vec::Vec<nika_check_analyzer::exec_read_path::Site>
+pub type nika_check_analyzer::exec_read_path::Site = (alloc::string::String, &'static str, alloc::string::String)
 pub mod nika_check_analyzer::native_first
 pub fn nika_check_analyzer::native_first::classify(command: &nika_schema::raw::action::RawCommand) -> core::option::Option<(&'static str, alloc::string::String)>
 pub fn nika_check_analyzer::native_first::classify_all(command: &nika_schema::raw::action::RawCommand) -> alloc::vec::Vec<(&'static str, alloc::string::String)>
+pub fn nika_check_analyzer::native_first::scan_sites(wf: &nika_schema::raw::workflow::RawWorkflow) -> alloc::vec::Vec<(alloc::string::String, &'static str, alloc::string::String)>
+pub mod nika_check_analyzer::schema_value
+pub fn nika_check_analyzer::schema_value::json_matches_type(value: &serde_json::value::Value, type_name: &str) -> bool
+pub fn nika_check_analyzer::schema_value::kind(v: &serde_json::value::Value) -> &'static str
 pub mod nika_check_analyzer::silent_literal
 pub fn nika_check_analyzer::silent_literal::scan(wf: &nika_schema::raw::workflow::RawWorkflow) -> alloc::vec::Vec<(alloc::string::String, alloc::string::String)>
 pub mod nika_check_analyzer::static_args
