+++ crates/nika-check-analyzer/src/lib.rs
@@ -44,11 +44,13 @@ mod builtin_shape;
 mod capacity;
 mod dag;
 pub mod edges;
+pub mod exec_read_path;
 mod jq_lint;
 pub mod native_first;
 mod scan;
 mod schema_lint;
 mod schema_paths;
+pub mod schema_value;
 pub mod silent_literal;
 pub mod static_args;
 mod static_ref;
